requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pandas==1.3.3
yfinance==0.1.63
httpx==0.18.2
pydantic==1.8.2 